order/requirements.txt
asgiref==3.6.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
pika==1.1.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
